requirements.txt
streamlit
yfinance
pandas
pytz
google-genai
python-dotenv
selenium
webdriver-manager
beautifulsoup4
matplotlib
finance-datareader